Revert "Merge branch 'next' into master"
This reverts commit 9eb2c8ef3605db8efb6e96eac043eb1d046514dd, reversing
changes made to e9ea2624dbc09c09a0bb656578c7888a739cf785.
