+++ package.json
@@ -54,8 +54,6 @@
     "@babel/core": "7.8.3",
     "@babel/polyfill": "7.8.3",
     "@babel/preset-typescript": "7.9.0",
-    "@emotion/core": "10.0.28",
-    "@emotion/styled": "10.0.27",
     "@testing-library/jest-dom": "^5.7.0",
     "@testing-library/react": "^10.0.4",
     "@types/chai": "4.1.7",
