+++ packages/gamut-labs/package.json
@@ -14,8 +14,6 @@
     "url": "git@github.com:Codecademy/client-modules.git"
   },
   "peerDependencies": {
-    "@emotion/core": ">=10.0.0",
-    "@emotion/styled": ">=10.0.0",
     "react": ">=16.8.1",
     "react-dom": ">=16.8.1"
   },
