+++ packages/gamut-labs/src/brand/Loading/index.tsx
@@ -1,4 +1,4 @@
-import { colors } from '@codecademy/gamut-styles';
+import { colors } from '@codecademy/gamut-styles/utils/variables';
 import cx from 'classnames';
 import React from 'react';
