+++ packages/gamut-styles/.gitignore
@@ -1,2 +1,1 @@
 *.d.ts
-dist
