+++ packages/gamut-styles/package.json
@@ -3,21 +3,14 @@
   "description": "Styleguide & Component library for codecademy.com",
   "version": "4.2.0",
   "author": "Jake Hiller <jake@codecademy.com>",
-  "main": "dist/index.js",
-  "module": "dist/index.js",
+  "main": "utils.scss",
   "scripts": {
-    "build:clean": "rm -rf ./dist",
-    "build:compile": "tsc",
-    "build": "yarn build:clean && yarn build:compile",
+    "build": "tsc",
     "prepublishOnly": "yarn build"
   },
   "repository": {
     "type": "git",
     "url": "git@github.com:RyzacInc/client-modules.git"
-  },
-  "peerDependencies": {
-    "@emotion/core": ">=10.0.0",
-    "@emotion/styled": ">=10.0.0"
   },
   "license": "MIT"
 }
