+++ packages/gamut/package.json
@@ -15,8 +15,6 @@
     "url": "git@github.com:RyzacInc/client-modules.git"
   },
   "peerDependencies": {
-    "@emotion/core": ">=10.0.0",
-    "@emotion/styled": ">=10.0.0",
     "react": ">=16.8.1",
     "react-dom": ">=16.8.1"
   },
