+++ packages/gamut/src/Badge/__tests__/Badge-test.tsx
@@ -4,10 +4,12 @@ import React from 'react';
 import Badge from '..';
 
 describe('Badge', () => {
-  it('renders badge text', () => {
-    const badgeText = 'I am a badge';
-    const wrapped = mount(<Badge>{badgeText}</Badge>);
+  it('combines classNames on its root when rootProps contains a className', () => {
+    const className = 'class-name';
+    const wrapped = mount(<Badge rootProps={{ className }} />);
 
-    expect(wrapped.text()).toEqual(badgeText);
+    const rootClassName = wrapped.childAt(0).prop('className');
+
+    expect(rootClassName).toMatch(new RegExp(`(.+) ${className}`));
   });
 });
