+++ packages/gamut/src/Badge/styles.module.scss
@@ -1,0 +1,11 @@
+@import "~@codecademy/gamut-styles/utils";
+
+.badge {
+  background: $color-blue-500;
+  border-radius: 3px;
+  color: $color-white;
+  display: inline-block;
+  font-size: 0.85em;
+  margin: 0 0.5em;
+  padding: 0.25em 0.75em;
+}
