+++ packages/styleguide/.storybook/main.js
@@ -18,10 +18,9 @@ module.exports = {
 
   addons: [
     '@storybook/addon-essentials',
+    '@storybook/addon-actions',
     '@storybook/addon-a11y',
     '@storybook/addon-docs',
-    '@storybook/addon-controls',
-    '@storybook/addon-actions',
     '@storybook/addon-links',
     '@storybook/addon-knobs/register',
   ],
