+++ packages/styleguide/package.json
@@ -23,17 +23,16 @@
     "@codecademy/gamut-icons": "^3.0.0",
     "@codecademy/gamut-styles": "^4.2.0",
     "@codecademy/webpack-config": "^5.0.2",
-    "@storybook/addon-a11y": "v6.0.0-rc.14",
-    "@storybook/addon-actions": "v6.0.0-rc.14",
-    "@storybook/addon-controls": "v6.0.0-rc.14",
-    "@storybook/addon-docs": "v6.0.0-rc.14",
-    "@storybook/addon-essentials": "v6.0.0-rc.14",
-    "@storybook/addon-knobs": "v6.0.0-rc.14",
-    "@storybook/addon-links": "v6.0.0-rc.14",
-    "@storybook/addon-storyshots": "v6.0.0-rc.14",
-    "@storybook/addons": "v6.0.0-rc.14",
-    "@storybook/react": "v6.0.0-rc.14",
-    "@storybook/theming": "v6.0.0-rc.14",
+    "@storybook/addon-a11y": "6.0.0-beta.39",
+    "@storybook/addon-actions": "6.0.0-beta.39",
+    "@storybook/addon-docs": "6.0.0-beta.39",
+    "@storybook/addon-essentials": "6.0.0-beta.39",
+    "@storybook/addon-knobs": "6.0.0-beta.39",
+    "@storybook/addon-links": "6.0.0-beta.39",
+    "@storybook/addon-storyshots": "6.0.0-beta.39",
+    "@storybook/addons": "6.0.0-beta.39",
+    "@storybook/react": "6.0.0-beta.39",
+    "@storybook/theming": "6.0.0-beta.39",
     "@types/react": "*",
     "@types/react-color": "^3.0.1",
     "autoprefixer": "^9.7.4",
