+++ packages/styleguide/stories/Core/Atoms/ProgressBar/ProgressBar.stories.mdx
@@ -2,7 +2,7 @@ import { Meta, Props, Preview, Story } from '@storybook/addon-docs/dist/blocks';
 import { StoryStatus } from '~styleguide/blocks';
 import { ProgressBar, LayoutGrid, Column } from '@codecademy/gamut/src';
 import { number } from '@storybook/addon-knobs';
-import { colors } from '@codecademy/gamut-styles';
+import { colors } from '@codecademy/gamut-styles/utils/variables';
 import { bars } from './constants';
 
 <Meta title="Atoms/ProgressBar" component={ProgressBar} />
