+++ packages/styleguide/stories/Core/Foundations/Colors/Colors.stories.mdx
@@ -23,7 +23,7 @@ import {
   editorColors,
   deprecatedGamutColors,
   platformColors,
-} from '@codecademy/gamut-styles';
+} from '@codecademy/gamut-styles/utils/variables';
 import { excludedColors, baseColors } from './constants';
 
 <Meta title="Foundations/Colors" component="Colors" />
