+++ packages/styleguide/stories/Core/Foundations/Typography/Typography.stories.mdx
@@ -9,7 +9,7 @@ import {
 } from '@codecademy/gamut/src';
 import { headingSizes, textSizes } from './constants';
 
-import { colors } from '@codecademy/gamut-styles';
+import { colors } from '@codecademy/gamut-styles/utils/variables';
 
 <Meta
   title="Foundations/Typography"
