+++ packages/styleguide/stories/Core/Layouts/Elements/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Item } from '@codecademy/gamut/src';
-import { colors } from '@codecademy/gamut-styles';
+import { colors } from '@codecademy/gamut-styles/utils/variables';
 
 export const Box: React.FC = ({ children }) => (
   <Item
